# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Onyx Meridian Enterprise OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /business-units/{id}:
    get:
      operationId: getBusinessUnit
      tags: [business-units]
      summary: Get a single business unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Business unit detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessUnitDetail"
        "404":
          description: Not found
  /business-units/{id}/kpis:
    get:
      operationId: getBusinessUnitKpis
      tags: [business-units]
      summary: Get KPIs for a business unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: KPI list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Kpi"
  /business-units/{id}/workflows:
    get:
      operationId: getBusinessUnitWorkflows
      tags: [business-units]
      summary: Get workflows for a business unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workflow list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
...
components:
  schemas:
...
    BusinessUnitDetail:
      type: object
      required: [id, name, domain, eeiScore, healthScore, activeKpis, activeAgents, activeWorkflows, status, risks, opportunities, cost, roi, forecastAccuracy, recentInsights]
      properties:
        id:
          type: string
        name:
          type: string
        domain:
          type: string
        eeiScore:
          type: number
        eeiTrend:
          type: number
        healthScore:
          type: number
        activeKpis:
          type: integer
        activeAgents:
          type: integer
        activeWorkflows:
          type: integer
        status:
          type: string
          enum: [healthy, at-risk, critical, optimizing]
        risks:
          type: integer
        opportunities:
          type: integer
        cost:
          type: number
        roi:
          type: number
        forecastAccuracy:
          type: number
        recentInsights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"
        eeiHistory:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
    Kpi:
      type: object
      required: [id, name, value, unit, target, status, trend, businessUnitId]
      properties:
        id:
          type: string
        name:
          type: string
        value:
          type: number
        unit:
          type: string
        target:
          type: number
        status:
          type: string
          enum: [on-track, watch, critical, surplus]
        trend:
          type: number
        businessUnitId:
          type: string
        owner:
          type: string
        linkedGoal:
          type: string
        description:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
    TrendPoint:
      type: object
      required: [date, value]
      properties:
        date:
          type: string
        value:
          type: number
    Workflow:
      type: object
      required: [id, name, status, businessUnitId, startedAt, stepCount, completedSteps, priority]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [running, paused, blocked, completed, failed]
        businessUnitId:
          type: string
        startedAt:
          type: string
        stepCount:
          type: integer
        completedSteps:
          type: integer
        priority:
          type: string
          enum: [p1, p2, p3]
        assignedAgent:
          type: ["string", "null"]
        blockers:
          type: ["string", "null"]
        estimatedCompletion:
          type: ["string", "null"]
...
    Insight:
      type: object
      required: [id, type, title, description, confidence, businessUnitId, priority, impact]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [prediction, recommendation, risk, opportunity, anomaly]
        title:
          type: string
        description:
          type: string
        confidence:
          type: number
        businessUnitId:
          type: string
        priority:
          type: string
          enum: [p1, p2, p3]
        impact:
          type: string
        evidence:
          type: array
          items:
            type: string
        projectedTrajectory:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        linkedKpi:
          type: ["string", "null"]
        createdAt:
          type: string
...