# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onyx Meridian Enterprise OS API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  date: string;
  value: number;
}